Reorganize toolbars to hide terms and conditions
